requirements.txt
appnope==0.1.0
backcall==0.1.0
blessings==1.6.1
certifi==2018.4.16
chardet==3.0.4
decorator==4.3.0
future==0.16.0
html5lib==1.0.1
httplib2==0.11.3
idna==2.6
inquirer==2.2.0
ipython==6.3.1
ipython-genutils==0.2.0
jedi==0.12.0
Jinja2==2.10
MarkupSafe==1.0
parso==0.2.0
pexpect==4.5.0
pickleshare==0.7.4
Pillow==5.1.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pycryptodomex==3.6.1
Pygments==2.2.0
pyjwkest==1.4.0
PyPDF2==1.26.0
python-dateutil==2.7.2
python-taiga==0.9.0
readchar==0.7
reportlab==3.4.0
requests==2.18.4
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
xhtml2pdf==0.2.2
